Move queries around to prevent dependency cycle
